add snackbar padding when meeting bar is present
